# store pointer to parent in TapeNodes instead of pushing to a tape stack