@jmxpearson 2016-07-01
Cool!

The other advantage of the graph is that one can query it, perform partial evaluations, etc. My understanding is that this is the Theano approach. 

But I was always worried about memory. 

Probably with Array nodes, this becomes much less of a concern.