@mlubin 2016-07-02
Yep. Note you can compare with JuMP's AD:
```jl
using JuMP
m = Model()
@variable(m, x[1:100000])
@NLobjective(m, Min, sum{ (1 - x[i])^2 + 100*(x[i+1]-x[i]^2)^2, i=1:length(x)-1})
d = JuMP.NLPEvaluator(m)
MathProgBase.initialize(d, [:Grad])
∇f = zeros(100000)
values = rand(100000)
@time MathProgBase.eval_grad_f(d, ∇f, values)
@time MathProgBase.eval_grad_f(d, ∇f, values)
```

On the second call I get:
```
  0.012159 seconds (9 allocations: 240 bytes)

```