@jrevels 2016-07-03
Wow, that's fast. 

Is that time including the full construction/destruction of the graph (i.e. forward pass + back-propagation + GC'ing the graph afterwards)? It doesn't seem like it from the memory measurements. 

Are you caching the graph for subsequent calls, and just reinterpreting it with new values? I don't think we can do that here, since our forward trace needs to accommodate value-dependent branches. 

I was going to ask about how to make our "interpreter" faster, it seems to partly be a game of what to share with the compiler and what to hide from it.