@mlubin 2016-07-03
The graph is stored between calls and we just do another pass through it
when the values change, so there's no GC involved for evaluating
derivatives.
On Jul 3, 2016 11:52 AM, "Jarrett Revels" <notifications@github.com> wrote:

> Wow, that's fast.
>
> Is that time including the full construction/destruction of the graph
> (i.e. forward pass + back-propagation + GC'ing the graph afterwards)? It
> doesn't seem like it from the memory measurements.
>
> Are you caching the graph for subsequent calls, and just reinterpreting it
> with new values? I don't think we can do that here, since our forward trace
> needs to accommodate value-dependent branches.
>
> I was going to ask about how to make our "interpreter" faster, it seems to
> partly be a game of what to share with the compiler and what to hide from
> it.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/jrevels/ReverseDiffPrototype.jl/pull/7#issuecomment-230157234>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/ABp0My6PaDLfLu_9b0A-M1cNB6FKzDodks5qR8yQgaJpZM4JCsLk>
> .
>